third_party/blink/web_tests/external/wpt/fetch/private-network-access/window-open.tentative.https.window-expected.txt

This is a testharness.js-based test.
[PASS] local to local: no preflight required.
[PASS] local to private: no preflight required.
[PASS] local to public: no preflight required.
[FAIL] private to local: failed preflight.
  assert_equals: expected "failure" but got "success"
[FAIL] private to local: missing CORS headers.
  assert_equals: expected "failure" but got "success"
[FAIL] private to local: missing PNA header.
  assert_equals: expected "failure" but got "success"
[PASS] private to local: success.
[PASS] private to private: no preflight required.
[PASS] private to public: no preflight required.
[FAIL] public to local: failed preflight.
  assert_equals: expected "failure" but got "success"
[FAIL] public to local: missing CORS headers.
  assert_equals: expected "failure" but got "success"
[FAIL] public to local: missing PNA header.
  assert_equals: expected "failure" but got "success"
[PASS] public to local: success.
[FAIL] public to private: failed preflight.
  assert_equals: expected "failure" but got "success"
[FAIL] public to private: missing CORS headers.
  assert_equals: expected "failure" but got "success"
[FAIL] public to private: missing PNA header.
  assert_equals: expected "failure" but got "success"
[PASS] public to private: success.
[PASS] public to public: no preflight required.
[FAIL] treat-as-public-address to local: failed preflight.
  assert_equals: expected "failure" but got "success"
[FAIL] treat-as-public-address to local: missing CORS headers.
  assert_equals: expected "failure" but got "success"
[FAIL] treat-as-public-address to local: missing PNA header.
  assert_equals: expected "failure" but got "success"
[PASS] treat-as-public-address to local: success.
[PASS] treat-as-public-address to local (same-origin): no preflight required.
[FAIL] treat-as-public-address to private: failed preflight.
  assert_equals: expected "failure" but got "success"
[FAIL] treat-as-public-address to private: missing CORS headers.
  assert_equals: expected "failure" but got "success"
[FAIL] treat-as-public-address to private: missing PNA header.
  assert_equals: expected "failure" but got "success"
[PASS] treat-as-public-address to private: success.
[PASS] treat-as-public-address to public: no preflight required.
[PASS] treat-as-public-address to local: optional preflight
Harness: the test ran to completion.
